+++ example/index.dart
@@ -2,14 +2,9 @@ import 'dart:html';
 
 void main() {
   InputElement tokenInput = querySelector('#token');
-  ButtonElement removeTokenBtn = querySelector('#removeTokenBtn');
-  String token = window.localStorage['token'];
+  String token = window.sessionStorage['token'];
   tokenInput.value = token;
   tokenInput.onKeyUp.listen((_) {
-    window.localStorage['token'] = tokenInput.value;
-  });
-  removeTokenBtn.onClick.listen((_) {
-    window.localStorage.remove('token');
-    tokenInput.value = '';
+    window.sessionStorage['token'] = tokenInput.value;
   });
 }